essay-genius-backend/requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langserve>=0.3.0
fastapi>=0.110.0
uvicorn>=0.28.0
python-dotenv>=1.0.0
pydantic>=2.7.0
sse_starlette>=1.6.5 